src/app/comment-drop/comment-drop.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { CommentDropComponent } from './comment-drop.component';

describe('Component: CommentDrop', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [CommentDropComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([CommentDropComponent],
      (component: CommentDropComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(CommentDropComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(CommentDropComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-comment-drop></app-comment-drop>
  `,
  directives: [CommentDropComponent]
})
class CommentDropComponentTestController {
}

